requirements.txt
streamlit
openai
python-dotenv
pandas
PyPDF2
openpyxl
requests
googlesearch-python